requirements.txt
asgiref==3.8.1
Django==5.1.2
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
wheel==0.44.0
whitenoise==6.8.2
